ADmin/CMS

+++ application/controllers/admin/Partners.php
@@ -20,20 +20,24 @@ class Partners extends Admin_Controller {
         $this->data['enable_editor'] = TRUE;
         $this->data['pageView'] = ADMIN . '/partners';
         
-        if (($_FILES["image"]["name"] != "")) {
-            $this->remove_file($this->uri->segment(4));
-            $image = upload_image(UPLOAD_PATH.'/partners', 'image');
-            generate_thumb(UPLOAD_PATH.'partners/',UPLOAD_PATH.'partners/',$image['file_name'],600,'thumb_');
-            if (!empty($image['file_name'])) {
-                $vals['image'] = $image['file_name'];
-            } else {
-                setMsg('error', 'Please upload a valid image file >> ' . strip_tags($image['error']));
-                redirect(ADMIN . '/partners/manage/' . $this->uri->segment(4), 'refresh');
+        if ($this->input->post()) {
+            $vals = $this->input->post();
+            if (($_FILES["image"]["name"] != "")) {
+                $this->remove_file($this->uri->segment(4));
+                $image = upload_image(UPLOAD_PATH.'/partners', 'image');
+                generate_thumb(UPLOAD_PATH.'partners/',UPLOAD_PATH.'partners/',$image['file_name'],600,'thumb_');
+                if (!empty($image['file_name'])) {
+                    $vals['image'] = $image['file_name'];
+                } else {
+                    setMsg('error', 'Please upload a valid image file >> ' . strip_tags($image['error']));
+                    redirect(ADMIN . '/partners/manage/' . $this->uri->segment(4), 'refresh');
+                }
+               
             }
             $this->master->insert_data('partners',$vals,'id',$this->uri->segment(4));
             setMsg('success', 'Partners has been saved successfully.');
             redirect(ADMIN . '/partners', 'refresh');
-        }
+        }    
         $this->data['row'] = $this->master->get_data_row('partners',array('id'=>$this->uri->segment(4)));
         $this->load->view(ADMIN . '/includes/siteMaster', $this->data);
     }
